Added the all-singing, all-dancing install-openstack-project script.
Just run it as root and give it a list of projects on the command line,
and it should take care of everything.  Verbosely shows what it's doing,
stops midway in the event of any errors.  Creds files that have been
processed and copied into users' home directorys are put into the
/root/ncicloudusers/DONE directory, so consider rm -rf
/root/ncicloudusers/DONE when you're sure everything is okay.